backend/ssj_project/requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
python-decouple
django-allauth
dj-rest-auth
django-filter
Pillow
requests
pyjwt
cryptography
gunicorn
whitenoise
dj-database-url
python-dotenv
msal
msgraph-core
djangorestframework-simplejwt